All warnings are solved with the exception of "Warning 4100: unreferenced formal parameter". They won't be solved because of the inheritance applied to all children from WinAPI class. Each control might handle the passing parameters according to its logic, so it's not wise to prune the parameter passing just to handle the warning. In addition to this, the project now treats Warning AS ERRORS (Level 4)! However, it's being compiled with the "/wd4100" advanced option to discard warning 4100 and satisfy our condition :D

+++ Windows-Wrapper/ListBox.h
@@ -21,10 +21,10 @@ private:
 	DockStyle m_DockStyle;
 	BorderStyle m_BorderStyle;
 	std::vector<RECT> m_RowPosition;
-	size_t m_TotalItemsInDrawableArea;
-	size_t m_ColumnWidth;
-	size_t m_RowNumber;
-	size_t m_ColumnNumber;
+	int m_TotalItemsInDrawableArea;
+	int m_ColumnWidth;
+	int m_RowNumber;
+	int m_ColumnNumber;
 	ListItem* m_SelectedItem;
 	std::vector<int> m_SelectedIndices;
 	std::vector<ListItem*> m_SelectedItems;
@@ -49,8 +49,8 @@ public:
 	bool IsMultiColumn() const noexcept;
 	void EnableMultiColumn() noexcept;
 	void DisableMultiColumn() noexcept;
-	size_t GetColumnWidth() noexcept;
-	void SetColumnWidth(const size_t& width) noexcept;
+	int GetColumnWidth() noexcept;
+	void SetColumnWidth(const int& width) noexcept;
 	BorderStyle GetBorderStyle() const noexcept;
 	void SetBorderStyle(BorderStyle style) noexcept;
 	SelectionMode GetSelectionMode() const noexcept;
